+++ autolog-core/src/main/java/com/github/maximevw/autolog/core/configuration/MethodInputLoggingConfiguration.java
@@ -166,10 +166,19 @@ public class MethodInputLoggingConfiguration {
 
 	/**
 	 * The logger name to use. If not specified, the default value {@value LoggingUtils#AUTOLOG_DEFAULT_TOPIC} will be
-	 * used.
+	 * used. To use the caller class name as logger name, keep this {@code null} or blank and set the property
+	 * {@link #isCallerClassUsedAsTopic()} to {@code true}.
 	 */
 	@API(status = API.Status.STABLE, since = "1.2.0")
 	private String topic;
+
+	/**
+	 * Whether the caller class name must be used as logger name when no custom logger name is specified in the property
+	 * {@link #getTopic()} ({@code null} or blank). By default: {@code false}.
+	 */
+	@API(status = API.Status.STABLE, since = "1.2.0")
+	@Builder.Default
+	private boolean callerClassUsedAsTopic = false;
 
     /**
      * Builds a new instance of configuration for auto-logging of input data of methods calls based on an annotation
@@ -193,6 +202,7 @@ public class MethodInputLoggingConfiguration {
 				.structuredMessage(autoLogMethodInOut.structuredMessage())
 				.dataLoggedInContext(autoLogMethodInOut.logDataInContext())
 				.topic(autoLogMethodInOut.topic())
+				.callerClassUsedAsTopic(autoLogMethodInOut.callerClassAsTopic())
                 .build();
     }
 
@@ -216,6 +226,7 @@ public class MethodInputLoggingConfiguration {
 				.structuredMessage(autoLogMethodInput.structuredMessage())
 				.dataLoggedInContext(autoLogMethodInput.logDataInContext())
 				.topic(autoLogMethodInput.topic())
+				.callerClassUsedAsTopic(autoLogMethodInput.callerClassAsTopic())
                 .build();
     }
 }