+++ autolog-core/src/main/java/com/github/maximevw/autolog/core/configuration/MethodOutputLoggingConfiguration.java
@@ -159,10 +159,19 @@ public class MethodOutputLoggingConfiguration {
 
 	/**
 	 * The logger name to use. If not specified, the default value {@value LoggingUtils#AUTOLOG_DEFAULT_TOPIC} will be
-	 * used.
+	 * used. To use the caller class name as logger name, keep this {@code null} or blank and set the property
+	 * {@link #isCallerClassUsedAsTopic()} to {@code true}.
 	 */
 	@API(status = API.Status.STABLE, since = "1.2.0")
 	private String topic;
+
+	/**
+	 * Whether the caller class name must be used as logger name when no custom logger name is specified in the property
+	 * {@link #getTopic()} ({@code null} or blank). By default: {@code false}.
+	 */
+	@API(status = API.Status.STABLE, since = "1.2.0")
+	@Builder.Default
+	private boolean callerClassUsedAsTopic = false;
 
     /**
      * Builds a new instance of configuration for auto-logging of output value of methods calls based on an annotation
@@ -191,6 +200,7 @@ public class MethodOutputLoggingConfiguration {
 				.structuredMessage(autoLogMethodInOut.structuredMessage())
 				.dataLoggedInContext(autoLogMethodInOut.logDataInContext())
 				.topic(autoLogMethodInOut.topic())
+				.callerClassUsedAsTopic(autoLogMethodInOut.callerClassAsTopic())
                 .build();
     }
 
@@ -213,6 +223,7 @@ public class MethodOutputLoggingConfiguration {
 				.structuredMessage(autoLogMethodOutput.structuredMessage())
 				.dataLoggedInContext(autoLogMethodOutput.logDataInContext())
 				.topic(autoLogMethodOutput.topic())
+				.callerClassUsedAsTopic(autoLogMethodOutput.callerClassAsTopic())
                 .build();
     }
 }